New fixture to avoid oauth config. Default admin.
This new fixture provides all the oauth configuration for the social
applications (client ID, secret key)..
For more information, please refer to:
[Fixtures Danjo](https://docs.djangoproject.com/en/1.10/howto/initial-data/)

This new fixture also provides a default admin called "admin". Thus, it is no
longer necessary to create a new superuser every time you clone the repository.
The password for the superuser is "adminadmin". For more information,
please refer to the README.md.