Implemented the view Upload.
If the user sends a GET request, this view will render a template
allowing the user to send a form.
If the user sends a POST request, this view will load the questions in
the database only if they are not already present. After loading all the
questions that can be loaded, the view renders a list of all the
questions that were saved and all the ones that weren't and why. If the
xml provided by the user is not valid the user will be warned and a
brief description explaining how the file should look like will be displayed
using xmlformathelp.html. If the question is too similar according to
the similarity function provided in similarity.py, then the question is
not saved and the user is warned.

Problems and some possible improvements:
- xmlformathelp is too simple and no specific information about the
  failure is given.
- similarity is not yet fully implemented, thus the similarity function
  will always return False. For now.