Implemented loading questions from xml files using admin.
admin was used as much as posible. A field in the list of xmlfiles was
added to allow the user to load questions from that file. Some files
were deleted to avoid code duplication and to avoid having two
interfaces that do exactly the same. Although the validation that was
done in the form.py is no longer possible, the lxml already check both
syntax and XSD validity. A consequence of this is that the user can
upload any file he or she wants, but he or she cannot parse it (he or
she can try it, but it will not work).

For more information in the implementation details, please refer to:
[adding actions](https://docs.djangoproject.com/en/1.10/ref/contrib/admin/actions/#adding-actions-to-the-modeladmin)
[Per-object custom actions ||| very good blog](http://morozov.ca/why-you-should-use-the-django-admin-9-tips.html)
[custom admin action](https://docs.djangoproject.com/en/1.10/ref/contrib/admin/actions/#custom-admin-action)