Now xml files are not saved in the db, they only load questions.
For some features to work the xml schema and the xml.py had to be rolled back.
When you upload a file, you are redirected to a different url where the
file is parsed and the questions are saved in the database. After that
the user is redirected to the list of questions, instead to the list of
xmlfiles Questions that are not saved because similar questions already exist,
are displayed in a warning message. Questions that are not saved
because identical questions already exist are displayed in an error
message. If the form is not valid, an error message is displayed and no
questions are saved.

messages.py is a small message handler. This was added to simplify the
admin.py implementation. I am not sure if this is the best way to do it.
I will appreciate comments and feedback about this.

forms.py was added with a few differences: now XMLFileForm is a
ModelForm and not a Form. This was done because we now have a model of
XMLFiles that we did not have previously.

models.py was changed in order to delete some fields in the XMLFile
model, in particular name and topic. The corresponding migration was added,
too.

xml.py was rolled back because it would otherwise break the code.

TODOs
-----

There are a few cases that must be implemented after commiting the new
xml schema:
- When the user uploads a question with a topic that does not exist,
  display an error message.
- When the user uploads a question with a subject that does not exist,
  display an error message.
- When the user uploads a question with both a topic and a subject that
  does not exist, display a (only one) error message.
Furthermore the following changes must be applied:
- Currently, the topic is selected choosing the first topic available.
  If no topic exists, the program will exit with error. So, after
  uploading the new xml schema the following items must be implemented:
  - The topic and subject must be extracted per question and used to
    determine the topic of the new question. This implies modifications
    in xml.py, admin.py and messages.py (load_questions_view function inside
    XMLFileAdmin)