Refactoring. Now the xmlfile content is validated.
Before, the file, questions and answers field lenghts were not
validated. This was a security bug, because someone could take the
server down by uploading a big file or a long question or answer (1).
Also, the server would let the user to upload similar or identical
questions if done using the GUI instead of the XML option (2).

In order to implement all these features, validation was used as
explained in:

Customizing the behaviour of clean() in the Question Model was necessary
to avoid the second issue (2). But this modification could be used to
have a better implementation of both parse_questions (xml.py) and
load_questions_view (admin.py). A few problems existed in the previous
implementations:
- parse_questions would not only retrieve the information but it would
  create instances of models. This violates the single responsability
  principle.
- load_questions_view had a very complex logic while it was doing a
  presumably simple task. This had to do with the fact that a lot a
  validation was done in the wrong place (it should have been carried
  out in the model).
All these problems were solved in this commit.

messages.py was also simplified because some validation error (that were not
instances of ValidationError)s were reduce to a single ValidationError.