Fix bug related to quiz.
After finishing the exam the site would still display all the elements.
This could not be exploited because the button would not allow the user
to send a post request again. Still, this was fixed and now a message is
displayed after the post is sent. This message is passed in context.

'sec' was deleted in order to make this html less language dependent.
Later on, if this is considered a requirement, we could add a trans
message.